+++ APIGateway-core/src/main/java/me/cxis/agw/dao/model/SysDO.java
@@ -10,7 +10,7 @@ public class SysDO {
 
     private Date modifiedTime;
 
-    private Short version;
+    private Integer version;
 
     private String name;
 
@@ -40,11 +40,11 @@ public class SysDO {
         this.modifiedTime = modifiedTime;
     }
 
-    public Short getVersion() {
+    public Integer getVersion() {
         return version;
     }
 
-    public void setVersion(Short version) {
+    public void setVersion(Integer version) {
         this.version = version;
     }
